requirements.txt
asgiref==3.3.1
Django==3.1.7
django-cors-headers==3.7.0
django-environ==0.4.5
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
mixins==0.1.4
psycopg2==2.8.6
PyJWT==2.0.1
pytz==2021.1
sqlparse==0.4.1
